+++ realm/realm-annotations-processor/src/main/java/io/realm/processor/Constants.java
@@ -27,7 +27,7 @@ public class Constants {
     public static final String TABLE_PREFIX = "class_";
     public static final String DEFAULT_MODULE_CLASS_NAME = "DefaultRealmModule";
     static final String STATEMENT_EXCEPTION_ILLEGAL_NULL_VALUE =
-            "throw new IllegalArgumentException(\"Trying to set non-nullable field %s to null.\")";
+            "throw new IllegalArgumentException(\"Trying to set non-nullable field '%s' to null.\")";
 
     static final Map<String, String> JAVA_TO_REALM_TYPES;
     static {