+++ realm/realm-library/src/main/java/io/realm/Realm.java
@@ -19,7 +19,6 @@ package io.realm;
 import android.annotation.TargetApi;
 import android.os.Build;
 import android.os.Looper;
-import android.os.Message;
 import android.util.JsonReader;
 
 import org.json.JSONArray;
@@ -790,12 +789,17 @@ public final class Realm extends BaseRealm {
      * you have a large number of object this method is generally faster.
      *
      * @param objects RealmObjects to insert.
+     * @throws IllegalStateException if the corresponding Realm is closed, called from an incorrect thread or not in a
+     * transaction.
      * @see #copyToRealm(Iterable)
      */
     public void insert(Collection<? extends RealmModel> objects) {
-        checkIfValid();
+        checkIfValidAndInTransaction();
         if (objects == null) {
             throw new IllegalArgumentException("Null objects cannot be inserted into Realm.");
+        }
+        if (objects.isEmpty()) {
+            return;
         }
         configuration.getSchemaMediator().insert(this, objects);
     }
@@ -816,10 +820,12 @@ public final class Realm extends BaseRealm {
      * you have a large number of object this method is generally faster.
      *
      * @param object RealmObjects to insert.
+     * @throws IllegalStateException if the corresponding Realm is closed, called from an incorrect thread or not in a
+     * transaction.
      * @see #copyToRealm(RealmModel)
      */
     public void insert(RealmModel object) {
-        checkIfValid();
+        checkIfValidAndInTransaction();
         if (object == null) {
             throw new IllegalArgumentException("Null object cannot be inserted into Realm.");
         }
@@ -843,12 +849,17 @@ public final class Realm extends BaseRealm {
      * you have a large number of object this method is generally faster.
      *
      * @param objects RealmObjects to insert.
+     * @throws IllegalStateException if the corresponding Realm is closed, called from an incorrect thread or not in a
+     * transaction.
      * @see #copyToRealmOrUpdate(Iterable)
      */
     public void insertOrUpdate(Collection<? extends RealmModel> objects) {
-        checkIfValid();
+        checkIfValidAndInTransaction();
         if (objects == null) {
             throw new IllegalArgumentException("Null objects cannot be inserted into Realm.");
+        }
+        if (objects.isEmpty()) {
+            return;
         }
         configuration.getSchemaMediator().insertOrUpdate(this, objects);
     }
@@ -869,10 +880,12 @@ public final class Realm extends BaseRealm {
      * you have a large number of object this method is generally faster.
      *
      * @param object RealmObjects to insert.
+     * @throws IllegalStateException if the corresponding Realm is closed, called from an incorrect thread or not in a
+     * transaction.
      * @see #copyToRealmOrUpdate(RealmModel)
      */
     public void insertOrUpdate(RealmModel object) {
-        checkIfValid();
+        checkIfValidAndInTransaction();
         if (object == null) {
             throw new IllegalArgumentException("Null object cannot be inserted into Realm.");
         }